requirements.txt
black==24.4.2
click==8.1.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pygame-ce==2.4.1
